Update URL tests
The current URL tests were not up to date with the latest URL spec, especially with regard to file URLs. In addition, they did not test serialization against predetermined results, but instead implemented an incorrect version of the URL Standard's algorithm (this is #2038). Finally, they left a lot of things implicit.

This moves to a new test data format that is explicit for every component of the URL, for maximum coverage. This includes previously-untested components like origin, username, and password.